requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
anthropic>=0.40.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
